+++ GymLedgerAPI/GymLedgerAPI/Domain/Interfaces /IGenericRepo.cs
@@ -6,7 +6,7 @@ namespace GymLedgerAPI.Domain.Interfaces
     public interface IGenericRepo<T> where T: class
     {
         ICollection<T> GetAll();
-        T GetbyId(long id);
+        T GetbyId(int id);
         void Add(T obj);
         void Remove(T obj);
         void SaveChanges();
